backend/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-dotenv==0.17.1
openai==0.28
python-multipart==0.0.6
boto3==1.28.44
pydantic==2.4.2
typing-extensions==4.8.0
starlette==0.27.0
passlib[bcrypt]
python-jose[cryptography]
click==8.0.1
Flask==2.0.1
gunicorn==23.0.0
itsdangerous==2.0.1
MarkupSafe>=2.1.1
Jinja2==3.1.6
sqlalchemy
psycopg2-binary
#MarkupSafe==2.0.1
#Werkzeug==2.2.3

